state logic fixes
